requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
openpyxl==3.1.5
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
requests==2.32.3